[♻️refactor]: 중복 사용되는 IPlant 타입을 전역으로 분리 리팩토링
